@@ -1,9 +1,8 @@
-namespace MVCApp.Models
-{
-    public class ErrorViewModel
-    {
-        public string? RequestId { get; set; }
+namespace MvcApp.Models;
 
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
+public class ErrorViewModel
+{
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 }
