@@ -1,9 +1,4 @@
-﻿using Dapper;
-using Microsoft.Extensions.Configuration;
-using System.Data;
-using System.Data.SqlClient;
-
-namespace DataLibrary.DataAccess;
+﻿namespace DataLibrary.DataAccess;
 
 public class SqlDataAccess : ISqlDataAccess
 {
