@@ -1,6 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace MVCApp.Models;
+﻿namespace MvcApp.Models;
 
 public class EmployeeDisplay
 {
