@@ -1,4 +1,4 @@
-﻿@model MVCApp.Models.EmployeeDisplay
+﻿@model MvcApp.Models.EmployeeDisplay
 
 @{
     ViewData["Title"] = "Employee Sign Up";
