@@ -1,4 +1,4 @@
-﻿@model IEnumerable<MVCApp.Models.EmployeeDisplay>
+﻿@model IEnumerable<MvcApp.Models.EmployeeDisplay>
 
 @{
     ViewData["Title"] = "View Employees";
@@ -43,13 +43,12 @@
                 <td>
                     @Html.DisplayFor(modelItem => item.EmailAddress)
                 </td>
-                @*
-            <td>
-            @Html.ActionLink("Edit", "Edit", new { /* id=item.PrimaryKey */ }) |
-            @Html.ActionLink("Details", "Details", new { /* id=item.PrimaryKey */ }) |
-            @Html.ActionLink("Delete", "Delete", new { /* id=item.PrimaryKey */ })
-            </td>
-            *@
+
+                <td>
+                    @Html.ActionLink("Details", "Details", new { /* id=item.PrimaryKey */ }) |
+                    @Html.ActionLink("Edit", "Edit", new { /* id=item.PrimaryKey */ }) |
+                    @Html.ActionLink("Delete", "Delete", new { /* id=item.PrimaryKey */ })
+                </td>
             </tr>
         }
     </tbody>
