@@ -1,3 +1,3 @@
-﻿@using MVCApp
-@using MVCApp.Models
+﻿@using MvcApp
+@using MvcApp.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
